Fixes CompatibilityAccessManager's receipt migration.
Since Purchases' restoreTransactions method may prompt for Apple ID and password as of version 3.6.0, this changes the restore method from restoreTransactions to syncPurchases (new in SDK 3.8.0). syncPurchases is silent and won't force refresh the receipt, and is useful when wanting to sync data with RevenueCat. The side effect is that the receipt may not be up-to-date, and thus may not sync all receipt data. To get the full up-to-date receipt, call restoreTransactions from a user-initiated action.